+++ plugin/src/main/groovy/org/openbakery/XcodeBuildPluginExtension.groovy
@@ -54,6 +54,7 @@ class XcodeBuildPluginExtension {
 	List<String> arch = null
 	String workspace = null
 	String version = null
+	Map<String, String> environment = null
 
 	boolean isOSX = false;
 	Devices devices = Devices.UNIVERSAL;
@@ -91,6 +92,17 @@ class XcodeBuildPluginExtension {
 			return project.getFileResolver().withBaseDir(project.getBuildDir()).resolve("shared")
 		}
 
+	}
+
+	String getWorkspace() {
+		if (workspace != null) {
+			return workspace
+		}
+		String[] fileList = project.projectDir.list(new SuffixFileFilter(".xcworkspace"))
+		if (fileList.length) {
+			return fileList[0]
+		}
+		return null
 	}
 
 	void setDstRoot(File dstRoot) {
@@ -252,6 +264,27 @@ class XcodeBuildPluginExtension {
 		}
 	}
 
+	void setEnvironment(Object environment) {
+		if (environment == null) {
+			return
+		}
+
+		if (environment instanceof Map) {
+			logger.debug("environment is Map: " + environment + " - " + environment.getClass().getName())
+			this.index = index;
+		} else {
+			logger.debug("environment is string: " + environment + " - " + environment.getClass().getName())
+			this.environment = new HashMap<String, String>();
+
+			String environmentString = environment.toString()
+			int index = environmentString.indexOf("=")
+			if (index == -1) {
+				environment.put(environmentString, null)
+			} else {
+				environment.put(environmentString.substring(0, index),environmentString.substring(index + 1))
+			}
+		}
+	}
 
 	void createXcode5DeviceList() {