+++ plugin/src/main/groovy/org/openbakery/signing/ProvisioningCleanupTask.groovy
@@ -20,10 +20,8 @@ import org.openbakery.AbstractXcodeTask
 
 
 class ProvisioningCleanupTask extends AbstractXcodeTask {
-	ProvisioningProfileIdReader provisioningProfileIdReader
 
 	ProvisioningCleanupTask() {
-		provisioningProfileIdReader = new ProvisioningProfileIdReader()
 	}
 
 	@TaskAction