+++ plugin/src/main/groovy/org/openbakery/signing/ProvisioningProfileIdReader.groovy
@@ -15,27 +15,48 @@
  */
 package org.openbakery.signing
 
+import org.apache.commons.lang.time.DateUtils
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
+
+import java.text.DateFormat
+import java.text.SimpleDateFormat
 
 
 class ProvisioningProfileIdReader {
 
-	def readProvisioningProfileIdFromDestinationRoot(def destinationRoot) {
-		logger.debug("destinationRoot: {}", destinationRoot);
-		if (!destinationRoot.exists()) {
-			return
+	private static Logger logger = LoggerFactory.getLogger(ProvisioningProfileIdReader.class)
+
+
+	boolean checkExpired(String source) {
+		def matcher = source =~ "<key>ExpirationDate</key>\\s*\\n\\s*<date>(.*?)</date>"
+		if (matcher.find()) {
+			String[] dateParsePatterns = ["yyyy-MM-dd'T'HH:mm:ssX"];
+			Date date = DateUtils.parseDate(matcher[0][1], dateParsePatterns);
+			if (date.before(new Date())) {
+				DateFormat formatter = DateFormat.getDateTimeInstance(DateFormat.LONG, DateFormat.LONG, Locale.getDefault());
+				throw new IllegalArgumentException("The Provisioning Profile has expired on " + formatter.format(date) );
+			}
 		}
+	}
 
-		def fileList = destinationRoot.list(
-						[accept: {d, f -> f ==~ /.*mobileprovision/ }] as FilenameFilter
-		).toList()
 
-		if (fileList.size() > 0) {
-			def mobileprovisionContent = new File(destinationRoot, fileList[0]).text
+	String readProvisioningProfileUUID(def source) {
+		logger.debug("destinationRoot: {}", source);
+		if (!(source instanceof File)) {
+			source = new File(source.toString())
+		}
+		if (source.exists()) {
+			def mobileprovisionContent = source.text
+			checkExpired(mobileprovisionContent);
+
 			def matcher = mobileprovisionContent =~ "<key>UUID</key>\\s*\\n\\s*<string>(.*?)</string>"
-			def uuid = matcher[0][1]
-			return uuid;
+			if (matcher.find()) {
+				return matcher[0][1];
+			}
 		}
 		return null;
 	}
+
 
 }